+++ apps/desktop/src/renderer/screens/main/components/WorkspaceSidebar/WorkspaceSidebarHeader/index.ts
@@ -1,0 +1,1 @@
+export { WorkspaceSidebarHeader } from "./WorkspaceSidebarHeader";
